Support using render_html output hook
If markdown uses the render_html output hook, r functions will be wrapped with
a span of class functioncall.  In these cases create a link to a google query
for the function name, searching the ggplot2 and inside-r.org sites for
documentation on the functioninside-r.org sites for documentation on the
function.